DRY up helper functions in protractor tests.
