+++ kifi-backend/modules/shoebox/angular/test/e2e/library.spec.js
@@ -1,33 +1,12 @@
 describe('kifi angular library test suite', function () {
   'use strict';
 
+  var util = require('./util/util');
+
+
   //
   // Helper functions.
   //
-  function loginUserWithCookies() {
-    // Set cookies to authenticate user.
-    // See: https://github.com/kifi/keepit/blob/master/integration/auth_headers.js
-    browser.manage().addCookie('KIFI_SECURESOCIAL', '567b766f-cae2-4bb9-b8f2-02e3ecb5cedd', '/', '.kifi.com');
-    browser.manage().addCookie('KIFI_SESSION', 'bd4cee3dd5d4170c702ff599302e223a900a8c4d-fortytwo_user_id=8397', '/', '.kifi.com');
-  }
-
-  function checkJSErrors() {
-    // Check for JavaScript errors that are in the Kifi minified JS file.
-    browser.manage().logs().get('browser').then(function (browserLog) {
-      var kifiJsErrors = [];
-      browserLog.forEach(function (log) {
-        if (log.message.match(/dist\/lib\.min\.\d+\.js/)) {
-          kifiJsErrors.push(log);
-        }
-      });
-
-      expect(kifiJsErrors.length).toEqual(0);
-      if (kifiJsErrors.length > 0) {
-        console.log('log: ' + require('util').inspect(kifiJsErrors));
-      }
-    });
-  }
-
   function checkForCorrectTitle(title) {
     expect(browser.getTitle()).toEqual(title);
   }
@@ -48,11 +27,11 @@ describe('kifi angular library test suite', function () {
   describe('Kifi main library page for logged in users', function () {
     beforeEach(function () {
       browser.get('/integrator-aa/main');
-      loginUserWithCookies();
+      util.loginUserWithCookies();
     });
 
     afterEach(function () {
-      checkJSErrors();
+      util.checkJSErrors();
     });
 
     it('should have the correct title', function () {
@@ -71,11 +50,11 @@ describe('kifi angular library test suite', function () {
   describe('Kifi secret library page for logged in users', function () {
     beforeEach(function () {
       browser.get('/integrator-aa/secret');
-      loginUserWithCookies();
+      util.loginUserWithCookies();
     });
 
     afterEach(function () {
-      checkJSErrors();
+      util.checkJSErrors();
     });
 
     it('should have the correct title', function () {
@@ -94,11 +73,11 @@ describe('kifi angular library test suite', function () {
   describe('Kifi public library page for logged in users', function () {
     beforeEach(function () {
       browser.get('/integrator-aa/test-1');
-      loginUserWithCookies();
+      util.loginUserWithCookies();
     });
 
     afterEach(function () {
-      checkJSErrors();
+      util.checkJSErrors();
     });
 
     it('should have the correct title', function () {
@@ -117,11 +96,11 @@ describe('kifi angular library test suite', function () {
   describe('Kifi private library page for logged in users', function () {
     beforeEach(function () {
       browser.get('/integrator-aa/test-2');
-      loginUserWithCookies();
+      util.loginUserWithCookies();
     });
 
     afterEach(function () {
-      checkJSErrors();
+      util.checkJSErrors();
     });
 
     it('should have the correct title', function () {
@@ -140,11 +119,11 @@ describe('kifi angular library test suite', function () {
   describe('Kifi extension library page for logged in users', function () {
     beforeEach(function () {
       browser.get('/integrator-aa/test-3');
-      loginUserWithCookies();
+      util.loginUserWithCookies();
     });
 
     afterEach(function () {
-      checkJSErrors();
+      util.checkJSErrors();
     });
 
     it('should have the correct title', function () {
@@ -163,11 +142,11 @@ describe('kifi angular library test suite', function () {
   describe('Kifi following library page for logged in users', function () {
     beforeEach(function () {
       browser.get('/lydialaurenson/halloween');
-      loginUserWithCookies();
+      util.loginUserWithCookies();
     });
 
     afterEach(function () {
-      checkJSErrors();
+      util.checkJSErrors();
     });
 
     it('should have the correct title', function () {
